memory small table

+++ stwo_cairo_prover/crates/prover/src/components/memory/mod.rs
@@ -36,7 +36,7 @@ mod tests {
             let decoded_id = memory.address_to_id[addr.0 as usize].decode();
             match decoded_id {
                 MemoryValueId::F252(id) => {
-                    id_to_f252.add_m31(BaseField::from_u32_unchecked(id as u32));
+                    id_to_f252.add_m31(BaseField::from_u32_unchecked(id));
                 }
                 MemoryValueId::Small(_id) => {}
             }
